Server/src/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
diart==0.9.2
fastapi==0.135.1
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
future==1.0.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
keras==3.13.2
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistral-common[audio]>=1.9.0
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numba==0.64.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.0.0
protobuf==7.34.0
pyannote.audio==3.4.0
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
setuptools==69.5.1
silero-vad==6.2.1
six==1.17.0
SpeechRecognition==3.15.1
starlette==0.52.1
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
termcolor==3.3.0
tf_keras==2.20.1
tiktoken==0.12.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.3
transformers>=5.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
logging==0.4.9.6
